+++ core/spec/lib/calculated_adjustments_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Spree::Core::CalculatedAdjustments do
+describe Spree::CalculatedAdjustments do
   it "should add has_one :calculator relationship" do
     assert Spree::ShippingMethod.reflect_on_all_associations(:has_one).map(&:name).include?(:calculator)
   end
